requirements.txt
python-telegram-bot==20.3
youtube-transcript-api==0.6.2
openai==1.14.3
faiss-cpu==1.7.4
tiktoken==0.6.0
python-dotenv==1.0.1
deep-translator==1.11.4
langchain==0.1.17
langchain-openai==0.1.3
langchain-community==0.0.36